Refactor blog editing structure: update import path for NewBlogPage and remove obsolete files from write directory.
