+++ src/app/blogs/[id]/edit/page.tsx
@@ -1,4 +1,4 @@
-import NewBlogPage from '../../write/NewBlogPage';
+import NewBlogPage from '../../new/NewBlogPage';
 export default async function Page({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
   const editId = Number(id);
