+++ src/app/blogs/new/page.tsx
@@ -1,0 +1,5 @@
+import NewBlogPage from './NewBlogPage';
+
+export default function Page() {
+  return <NewBlogPage />;
+}
